+++ Configurator/src/utils/constants.ts
